Extract out Criteria visiting for Facets.
