+++ Source/ElasticLINQ/Request/Visitors/ElasticFieldsExpressionVisitor.cs
@@ -1,9 +1,9 @@
 ﻿// Licensed under the Apache 2.0 License. See LICENSE.txt in the project root for more information.
 
-using System;
 using ElasticLinq.Mapping;
 using ElasticLinq.Response.Model;
 using ElasticLinq.Utility;
+using System;
 using System.Linq.Expressions;
 
 namespace ElasticLinq.Request.Visitors
